# api/openapi/oiaf.openapi.yaml
openapi: 3.0.3
info:
  title: OIAF API
  description: >
    Open Identity Access Firewall (OIAF) control-plane API. Risk-based access
    control and MFA orchestration for hybrid identity.
  version: 0.1.0
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://127.0.0.1:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/access/evaluate:
    post:
      tags: [access]
      summary: Evaluate an access request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRequest"
      responses:
        "200":
          description: Access decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessDecision"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /v1/challenge/{id}/verify:
    post:
      tags: [challenge]
      summary: Verify a challenge (TOTP or push)
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
                  enum: [totp, push]
                code:
                  type: string
                device_id:
                  type: string
                number:
                  type: integer
                signature:
                  type: string
                timestamp:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/identities:
    get:
      tags: [identities]
      summary: List identities
      responses:
        "200":
          description: Identities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Identity"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [identities]
      summary: Create identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Identity"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Identity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/identities/{id}:
    get:
      tags: [identities]
      summary: Get identity
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Identity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [identities]
      summary: Update identity
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Identity"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Identity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [identities]
      summary: Delete identity
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/identities/{id}/factors/totp/enroll:
    post:
      tags: [identities]
      summary: Enroll a TOTP factor
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Enrollment details
          content:
            application/json:
              schema:
                type: object
                properties:
                  factor_id:
                    type: string
                  secret:
                    type: string
                  otpauth_uri:
                    type: string
                  status:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/identities/{id}/factors/totp/activate:
    post:
      tags: [identities]
      summary: Activate a TOTP factor
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                factor_id:
                  type: string
                code:
                  type: string
      responses:
        "200":
          description: Activated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/devices:
    get:
      tags: [devices]
      summary: List devices
      responses:
        "200":
          description: Devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [devices]
      summary: Register a push device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identity_id:
                  type: string
                name:
                  type: string
      responses:
        "201":
          description: Device and one-time secret
          content:
            application/json:
              schema:
                type: object
                properties:
                  device:
                    $ref: "#/components/schemas/Device"
                  secret:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/devices/{id}:
    get:
      tags: [devices]
      summary: Get device
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [devices]
      summary: Delete device
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/devices/{id}/approve-challenge:
    post:
      tags: [devices]
      summary: Approve a push challenge from a device
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                challenge_id:
                  type: string
                number:
                  type: integer
                signature:
                  type: string
                timestamp:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/policies:
    get:
      tags: [policies]
      summary: List policies
      responses:
        "200":
          description: Policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [policies]
      summary: Create policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Policy"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/policies/test:
    post:
      tags: [policies]
      summary: Test a policy against a request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                policy:
                  $ref: "#/components/schemas/Policy"
                request:
                  $ref: "#/components/schemas/AccessRequest"
      responses:
        "200":
          description: Policy decision
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/adapters/{id}:
    get:
      tags: [adapters]
      summary: Get adapter
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Adapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adapter"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [adapters]
      summary: Delete adapter
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/audit/verify:
    get:
      tags: [audit]
      summary: Verify the audit hash chain
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    AccessDecision:
      type: object
      properties:
        request_id:
          type: string
        decision:
          type: string
          enum: [allow, deny, challenge, alert]
        risk_score:
          type: integer
          minimum: 0
          maximum: 100
        reasons:
          type: array
          items:
            type: string
        challenge:
          type: object
          properties:
            id:
              type: string
            methods:
              type: array
              items:
                type: string
                enum: [totp, push, webauthn, email, sms]
            expires_at:
              type: string
              format: date-time
...